Introduce an optimization for "small" maps (raw capacity fits in 32 bits)
Pos { u64 } is interpreted dynamically.

An OrderedMap whose capacity fits in 32 bits interprets Pos as an u32
entry index and a u32 truncated hash value. Note that the 32 lower bits
of the hash are enough for computing desired position and probe distance
in this case.

An OrderedMap whose capacity needs at least 32 bits, interprets Pos as
an u64 entry index, and needs to retrieve the hash from the Entry struct.

Having both index and hash in the indices vector increases cache
utilization, and thus benefits lookup and other hash map operations.