FEAT: Make .retain() in-order and preserve-order
We want to visit the elements in order, and keep them in order after the
retain; the complicated part is removing them and backward shifting them
correctly.

The new version is a bit slower than HashMap::retain and than the old
unordered retain, so we keep around the unordered version too for those
that accept jumbling up the order.