FEAT: Add internal methods for slicing using RangeBounds
These allow us to slice using RangeBounds<usize> which includes all the
range types (a..b, c..=d) and some extras.

The main infidelity of the interface is the need to handle
Bound::Exclusive(std::usize::MAX) which does not have an equivalent in
the i..j or i..=j and slices can't handle it natively.

The scaffolding in src/util to create the .slice() and .slice_mut()
methods is somewhat verbose, but it allows us to write neat code where
it is used.