Be more careful about reserving entries capacity
Blindly following the indices capacity could make the entries exceed the
maximum layout size, `isize::MAX`, so now we use a limit against that.
Furthermore, even within that limit we can *try* the reservation first,
and fall back to the minimum requested growth if that fails.

This shouldn't matter in practice for 64-bit targets, but smaller
targets could potentially run close to the max allocation size.