Re-deprecate `remove`, `remove_entry`, and `take`
It's a frequent point of confusion that these methods disrupt the order
(by swapping the last item), but that choice was made for performance.
The alternative `shift_*` removals are also disruptive to indices, and
require O(n) updates instead of O(1).

The deprecation notes recommend instead choosing a swap/shift method
explicitly. This doesn't mean that we'll ever actually _remove_ these
deprecated methods, since they are still nice to have when acting as a
drop-in `HashMap` replacement, but the warning will hopefully help avoid
surprises.