Use the bucket API from hashbrown v0.16.1
This lets us simplify `OccupiedEntry` to just a reference and a bucket
index, rather than needing a separate reference to our `Vec` while the
inner entry holds a reference to the `HashTable`. This in turn lets us
get rid of the `RefMut` abstraction that was providing common methods
for that pair of references.

The `OccupiedEntry` currently holds the index for the vector as well,
but this is more of an optimization since most entry methods need it, so
we don't have to indirectly read it through the bucket each time. Thus
the overall size hasn't changed, but we could drop that if we decided
the size was more important.

The sizes of `VacantEntry` and `IndexedEntry` *are* reduced though,
because they're also now holding a single map reference rather than the
double `RefMut`, which was only needed before due the structure of the
`OccupiedEntry`. The overall `Entry` should also be smaller as a result.